Temporary bodge for grid issue
The countries grid is sampled at the interstices of a 500 x 250
lattice, so it has 501 x 251 points.

A separate bug meant the last row was omitted, which is fairly
harmless since it is blank, but contributed to my confusion.

The correct fix would be to resample at the centres of the lattice
squares; here we are just truncating the grid to the right dimensions,
which means it’s offset by half a unit in each direction.